refactor: move feerate type to bitcoin module
